+++ RandomForeseerCode/Common/PredictionExtensions.cs
@@ -1,0 +1,12 @@
+
+using MegaCrit.Sts2.Core.Runs;
+
+namespace RandomForeseer.RandomForeseerCode.Common;
+
+internal static class PredictionExtensions
+{
+    public static RelicGrabBag Clone(this RelicGrabBag grabBag)
+    {
+        return RelicGrabBag.FromSerializable(grabBag.ToSerializable());
+    }
+}